Added bases for Line primitive
Line primitive will be just a square with Special treatement. It's choosen this way, due to simplicity in mind. I've implemented rotation method, which will rotate the squares. For now I think it doesn't look as it should.

Updated main to show off rotation of squares, I think It looks cool at least.